refactor: split large files, improve mobile chart UX
Backend:
- Split repository.go into domain files (users, projects, apikeys, spans, aggregates, alerts)
- Extract CLI subcommands from main.go into cli.go
- Split service/query.go into query_services, query_traces, query_dependencies

Frontend:
- Reduce chart left padding on mobile, add duration p50/p95/p99 stats
- Add p99/p95/p50 legend to latency chart
- Make span attributes table scroll horizontally
- Tighter card padding on mobile
