+++ internal/repository/repo_users.go
@@ -1,0 +1,56 @@
+package repository
+
+import "database/sql"
+
+func (r *Repository) CreateUser(username, passwordHash string) error {
+	_, err := r.db.Exec("INSERT INTO users (username, password_hash) VALUES (?, ?)", username, passwordHash)
+	return err
+}
+
+func (r *Repository) GetUserByUsername(username string) (User, error) {
+	var u User
+	err := r.db.QueryRow("SELECT id, username, password_hash, created_at FROM users WHERE username = ?", username).
+		Scan(&u.ID, &u.Username, &u.PasswordHash, &u.CreatedAt)
+	return u, err
+}
+
+func (r *Repository) UpdateUserPassword(username, passwordHash string) error {
+	res, err := r.db.Exec("UPDATE users SET password_hash = ? WHERE username = ?", passwordHash, username)
+	if err != nil {
+		return err
+	}
+	n, _ := res.RowsAffected()
+	if n == 0 {
+		return sql.ErrNoRows
+	}
+	return nil
+}
+
+func (r *Repository) DeleteUser(username string) error {
+	res, err := r.db.Exec("DELETE FROM users WHERE username = ?", username)
+	if err != nil {
+		return err
+	}
+	n, _ := res.RowsAffected()
+	if n == 0 {
+		return sql.ErrNoRows
+	}
+	return nil
+}
+
+func (r *Repository) ListUsers() ([]User, error) {
+	rows, err := r.db.Query("SELECT id, username, password_hash, created_at FROM users ORDER BY id")
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var out []User
+	for rows.Next() {
+		var u User
+		if err := rows.Scan(&u.ID, &u.Username, &u.PasswordHash, &u.CreatedAt); err != nil {
+			return nil, err
+		}
+		out = append(out, u)
+	}
+	return out, rows.Err()
+}
