+++ web/src/pages/OperationDetailPage.tsx
@@ -11,6 +11,7 @@ import {
   XAxis,
   YAxis,
   Tooltip,
+  Legend,
   ResponsiveContainer,
   CartesianGrid,
 } from 'recharts'
@@ -60,6 +61,11 @@ export function OperationDetailPage(): ReactElement {
     errorRate: b.count > 0 ? (b.errorCount / b.count) * 100 : 0,
   }))
 
+  const totalCount = timeseries.reduce((s, b) => s + b.count, 0)
+  const avgP50 = totalCount > 0 ? timeseries.reduce((s, b) => s + b.p50Us * b.count, 0) / totalCount : 0
+  const avgP95 = totalCount > 0 ? timeseries.reduce((s, b) => s + b.p95Us * b.count, 0) / totalCount : 0
+  const avgP99 = totalCount > 0 ? timeseries.reduce((s, b) => s + b.p99Us * b.count, 0) / totalCount : 0
+
   if (loading) {
     return (
       <div>
@@ -96,6 +102,24 @@ export function OperationDetailPage(): ReactElement {
         <TimeRangeSelector value={range} onChange={setRange} />
       </div>
 
+      {/* Duration stats */}
+      <div style={{ display: 'grid', gridTemplateColumns: 'repeat(3, 1fr)', gap: '0.75rem', marginBottom: '1rem' }}>
+        {[
+          { label: 'p50', value: avgP50, color: '#3b82f6' },
+          { label: 'p95', value: avgP95, color: '#eab308' },
+          { label: 'p99', value: avgP99, color: '#ef4444' },
+        ].map((s) => (
+          <div key={s.label} className="card" style={{ textAlign: 'center' }}>
+            <div style={{ fontSize: '0.75rem', color: 'var(--text-muted)', textTransform: 'uppercase', letterSpacing: '0.05em', marginBottom: 4 }}>
+              {s.label}
+            </div>
+            <div style={{ fontSize: '1.25rem', fontWeight: 700, color: s.color, fontFamily: "'SF Mono', 'Fira Code', monospace" }}>
+              {formatDuration(s.value)}
+            </div>
+          </div>
+        ))}
+      </div>
+
       {/* Charts grid */}
       <div className="charts-grid" style={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: '1rem', marginBottom: '1.5rem' }}>
         {/* Latency chart */}
@@ -103,11 +127,11 @@ export function OperationDetailPage(): ReactElement {
           <div style={{ fontSize: '0.875rem', fontWeight: 600, marginBottom: '0.75rem' }}>
             Latency (ms)
           </div>
-          <ResponsiveContainer width="100%" height={180}>
-            <LineChart data={chartData}>
+          <ResponsiveContainer width="100%" height={200}>
+            <LineChart data={chartData} margin={{ top: 4, right: 8, bottom: 0, left: -16 }}>
               <CartesianGrid strokeDasharray="3 3" stroke="var(--border)" />
               <XAxis dataKey="time" tick={{ fontSize: 11, fill: 'var(--text-muted)' }} />
-              <YAxis tick={{ fontSize: 11, fill: 'var(--text-muted)' }} />
+              <YAxis tick={{ fontSize: 11, fill: 'var(--text-muted)' }} width={40} />
               <Tooltip
                 contentStyle={{
                   background: 'var(--surface)',
@@ -115,10 +139,17 @@ export function OperationDetailPage(): ReactElement {
                   borderRadius: 8,
                   color: 'var(--text)',
                 }}
+                formatter={(value: number) => [`${value.toFixed(2)} ms`]}
               />
-              <Line type="monotone" dataKey="p99" stroke="#ef4444" dot={false} strokeWidth={2} />
-              <Line type="monotone" dataKey="p95" stroke="#eab308" dot={false} strokeWidth={2} />
-              <Line type="monotone" dataKey="p50" stroke="#3b82f6" dot={false} strokeWidth={2} />
+              <Legend
+                verticalAlign="top"
+                align="right"
+                iconType="line"
+                wrapperStyle={{ fontSize: 11, paddingBottom: 4 }}
+              />
+              <Line type="monotone" dataKey="p99" name="p99" stroke="#ef4444" dot={false} strokeWidth={2} />
+              <Line type="monotone" dataKey="p95" name="p95" stroke="#eab308" dot={false} strokeWidth={2} />
+              <Line type="monotone" dataKey="p50" name="p50" stroke="#3b82f6" dot={false} strokeWidth={2} />
             </LineChart>
           </ResponsiveContainer>
         </div>
@@ -129,10 +160,10 @@ export function OperationDetailPage(): ReactElement {
             Throughput (requests)
           </div>
           <ResponsiveContainer width="100%" height={180}>
-            <BarChart data={chartData}>
+            <BarChart data={chartData} margin={{ top: 4, right: 8, bottom: 0, left: -16 }}>
               <CartesianGrid strokeDasharray="3 3" stroke="var(--border)" />
               <XAxis dataKey="time" tick={{ fontSize: 11, fill: 'var(--text-muted)' }} />
-              <YAxis tick={{ fontSize: 11, fill: 'var(--text-muted)' }} />
+              <YAxis tick={{ fontSize: 11, fill: 'var(--text-muted)' }} width={40} />
               <Tooltip
                 contentStyle={{
                   background: 'var(--surface)',
@@ -153,10 +184,10 @@ export function OperationDetailPage(): ReactElement {
           Error Rate (%)
         </div>
         <ResponsiveContainer width="100%" height={150}>
-          <AreaChart data={chartData}>
+          <AreaChart data={chartData} margin={{ top: 4, right: 8, bottom: 0, left: -16 }}>
             <CartesianGrid strokeDasharray="3 3" stroke="var(--border)" />
             <XAxis dataKey="time" tick={{ fontSize: 11, fill: 'var(--text-muted)' }} />
-            <YAxis tick={{ fontSize: 11, fill: 'var(--text-muted)' }} />
+            <YAxis tick={{ fontSize: 11, fill: 'var(--text-muted)' }} width={40} />
             <Tooltip
               contentStyle={{
                 background: 'var(--surface)',
